+++ lib/atprotocol/app/bsky/feed/appbskyfeedgetquotes.h
@@ -10,7 +10,8 @@ class AppBskyFeedGetQuotes : public AppBskyFeedGetPosts
 public:
     explicit AppBskyFeedGetQuotes(QObject *parent = nullptr);
 
-    void getQuotes(const QString &uri, const QString &cid, const int limit, const QString &cursor);
+    void getQuotes(const QString &uri, const QString &cid, const qint64 limit,
+                   const QString &cursor);
 };
 
 }
